+++ app/src/main/java/fr/levitt/profiling/data/ReverseHashRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReverseHashRepository extends CrudRepository<ReverseHash, Integer> {
+public interface ReverseHashRepository extends CrudRepository<ReverseHash, Long> {
 
 	List<ReverseHash> findByHash(String hash);
 }
